Update config
Too lazy to break some of these changes out.

A majority of the changes involve cleaning up config functions, and
adding in empty opts tables for setup functions that require opts to not
be nil.

Some of the changes are moving which module adds the plugin, i.e. moving
a majority of the plugins out of lua/plugins/init.lua and into more
appropriate locations.

There are also a few colorschemes that have been added.